requirements.txt
Django==4.2.30
djangorestframework==3.15.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
celery==5.4.0
redis==5.0.4
mysqlclient==2.2.4
python-dotenv==1.0.1
